% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MeanShift_Classical}
\alias{MeanShift_Classical}
\title{Mean shift clustering}
\usage{
MeanShift_Classical(
  pc,
  CD2TH_fac,
  CH2TH_fac,
  MaxIter = 200L,
  UniformKernel = FALSE
)
}
\arguments{
\item{pc}{Point cloud data in a three-column matrix where the columns
represent X, Y and Z coordinates and each row represents one point.}

\item{CD2TH_fac}{Numeric scalar. Ratio of crown diameter to tree height.
Determines kernel diameter based on the height of its center.}

\item{CH2TH_fac}{Numeric scalar. Ratio of crown height to tree height.
Determines kernel height based on the height of its center.}

\item{MaxIter}{Integer scalar. Maximum number of iterations, i.e. steps that
the kernel can move for each point. If no mode is found after
\code{MaxIter} iterations, the centroid that was calculated last is
treated as the mode.}

\item{UniformKernel}{Boolean scalar. Set to \code{TRUE} in order to turn off
distance weighting within the kernel (defaults to \code{FALSE}).}
}
\value{
A data.frame with the coordinates in \code{pc} and three additional
columns with the coordinates of the calculated modes.
}
\description{
Adaptive mean shift clustering to delineate tree crowns from lidar point
clouds.
}
