% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{meanShiftClassicImproved}
\alias{meanShiftClassicImproved}
\title{Mean shift clustering}
\usage{
meanShiftClassicImproved(
  pointCloud,
  crownDiameter2TreeHeight,
  crownHeight2TreeHeight,
  maxNumCentroidsPerMode = 200L
)
}
\arguments{
\item{pointCloud}{Point cloud data in a three-column matrix where the
columns represent X, Y and Z coordinates and each row represents one
point.}

\item{crownDiameter2TreeHeight}{Numeric scalar. Ratio of crown diameter
to tree height. Determines kernel diameter based on the height of its
center.}

\item{crownHeight2TreeHeight}{Numeric scalar. Ratio of crown height to tree
height. Determines kernel height based on the height of its center.}

\item{maxNumCentroidsPerMode}{Integer scalar. Maximum number of
iterations, i.e. steps that the kernel can move for each point. If no mode
is found after \code{maxNumCentroidsPerMode} iterations, the centroid
that was calculated last is treated as the mode.}
}
\value{
A data.frame with the coordinates in \code{pointCloud} and three
additional columns with the coordinates of the calculated modes.
}
\description{
Adaptive mean shift clustering to delineate tree crowns from lidar point
clouds.
}
