% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_mean_shift.R
\name{parallel_mean_shift}
\alias{parallel_mean_shift}
\title{Parallel mean shift clustering for individual tree crown delineation}
\usage{
parallel_mean_shift(
  point_clouds,
  used_fraction_of_cores = 0.5,
  version = "classic",
  crown_diameter_2_tree_height,
  crown_height_2_tree_height,
  max_num_centroids_per_mode = 200,
  buffer_width = 10,
  min_height = 2,
  mode_rounding = 2
)
}
\arguments{
\item{point_clouds}{List of point clouds in data.table format containing
columns X, Y and Z (produced by the \code{split_point_cloud_buffered}
function).}

\item{used_fraction_of_cores}{Fraction of available cores to use for
parallelization.}

\item{version}{of the AMS3D algorithm. Can be set to "classic" (slow but
precise also with small trees) or "voxel" (fast but based on rounded
coordinates of 1-m precision) or "classic improved" (like classic but
faster).}

\item{crown_diameter_2_tree_height}{Factor for the ratio of height to crown
width. Determines kernel diameter based on its height above ground.}

\item{crown_height_2_tree_height}{Factor for the ratio of height to crown
length. Determines kernel height based on its height above ground.}

\item{max_num_centroids_per_mode}{Maximum number of iterations, i.e. steps
that the kernel can move for each point. If centroid is not found after all
iteration, the last position is assigned as centroid and the processing
jumps to the next point}

\item{buffer_width}{Width of the buffer around the core area in meters.}

\item{min_height}{Minimum height above ground for a point to be considered in
the analysis. Has to be > 0.}

\item{mode_rounding}{Numeric Scalar. Specifies the rounding accuracy for mode
positions. After rounding all modes with the same coordinates are
considered to belong to one tree crown.}
}
\value{
data.table of point cloud with points labelled with tree IDs
}
\description{
The function provides the frame work to apply the adaptive mean shift 3D
(AMS3D) algorithm on several sub point clouds of a large investigation area
in parallel. It requires a list of sub point clouds as input and returns one
large clustered point cloud as output. The input should have buffer zones
around the focal areas. The buffer width should correspond to at least the
maximal possible tree crown radius.
}
