% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_tree_crowns.R
\name{segment_tree_crowns}
\alias{segment_tree_crowns}
\title{Calculate crown IDs for trees in a point cloud}
\usage{
segment_tree_crowns(
  point_cloud,
  version = "classic",
  crown_diameter_2_tree_height,
  crown_height_2_tree_height,
  max_num_centroids_per_mode = 200,
  min_num_neighbors_per_core,
  neighborhood_radius
)
}
\arguments{
\item{point_cloud}{A data.frame or data.table. Its first three columns are
expected to hold coordinates.}

\item{version}{Character. One of "classic" or "improved".}
}
\description{
Calculate crown IDs for trees in a point cloud
}
