% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_point_cloud_buffered.R
\name{split_point_cloud_buffered}
\alias{split_point_cloud_buffered}
\title{Split point cloud into subsets with buffer areas around them}
\usage{
split_point_cloud_buffered(point_cloud, core_width, buffer_width)
}
\arguments{
\item{point_cloud}{A data.table containing columns with x-, y-, and z-
coordinates.}

\item{core_width}{Width of the core area in meters.}

\item{buffer_width}{Width of the buffer around the core area in meters.}
}
\value{
List of data.tables that each contains the coordinates of a point
cloud subset together with a boolean column "Buffer" that labels core- and
buffer-points.
}
\description{
The function splits one large point cloud into several smaller point clouds.
It allows to specify a buffer width around the core area where points are
included.
}
